mobile lightbox
was previously creeping into the top of the poster moment on the page while on mobile, as it is stashed away above the poster, to slide in from the top when enabled. ensured that it is out of site until it is called upon to be in site.

mobile menu scrolls through list of departments if they go beyond the bottom of the viewport.