consistent filtering
filtering is now done in the same process tick as checking against the render conditionals. this way projects can not make it through the filter process, have the filter be changed, and that project still making it to the render function because it had already passed the filter when it was set to some other value.